move features/Wallets to services/WalletService
