+++ common/v2/components/WalletUnlock/Web3Provider.tsx
@@ -4,7 +4,7 @@ import translate, { translateRaw } from 'translations';
 import { KNOWLEDGE_BASE_URL as KB_URL } from 'v2/config';
 import MetamaskSVG from 'common/assets/images/wallets/metamask-2.svg';
 import { NewTabLink } from 'components/ui';
-import { unlockWeb3 } from 'v2/features/Wallets';
+import { unlockWeb3 } from 'v2/services/WalletService';
 import './Web3Provider.scss';
 import { InlineErrorMsg } from 'v2/components';
