+++ common/v2/services/WalletService/mnemonic/types.ts
